backend/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiosqlite==0.19.0",
    "fastapi==0.103.1",
    "httpx>=0.28.1",
    "jupyterlab>=4.4.2",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pillow==10.0.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.21",
    "uvicorn[standard]==0.23.2",
]
